add @login_required tags to superuser specific views.py

+++ products/views.py
@@ -1,8 +1,10 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Product, Category
 from .forms import ProductForm
+
 
 # Create your views here.
 
@@ -68,8 +70,13 @@ def product_detail(request, product_id):
     return render(request, 'product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Superusers can do this')
+        return redirect(reverse, 'home')
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -87,35 +94,43 @@ def add_product(request):
         'form': form,
     }
 
-    return render(request, template, context)
+        render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
-    product = get_object_or_404(Product, pk=product_id)
-    if request.method == 'POST':
-        form = ProductForm(request.POST, request.FILES, instance=product)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Successfully updated product!')
-            return redirect(reverse('product_detail', args=[product.id]))
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Superusers can do this')
+        return redirect(reverse, 'home')
+        product = get_object_or_404(Product, pk=product_id)
+        if request.method == 'POST':
+            form = ProductForm(request.POST, request.FILES, instance=product)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Successfully updated product!')
+                return redirect(reverse('product_detail', args=[product.id]))
+            else:
+                messages.error(request, 'Failed to update product, please try again.')
         else:
-            messages.error(request, 'Failed to update product, please try again.')
-    else:
-        form = ProductForm(instance=product)
-        messages.info(request, f'You are editing {product.name}')
+            form = ProductForm(instance=product)
+            messages.info(request, f'You are editing {product.name}')
 
-    template = 'update-product.html'
-    context = {
-        'form': form,
-        'product': product,
-    }
+        template = 'update-product.html'
+        context = {
+            'form': form,
+            'product': product,
+        }
 
-    return render(request, template, context)
+        return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete Product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Superusers can do this')
+        return redirect(reverse, 'home')
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product Deleted Successfully!')
